refactor: move constants in init file
