+++ packages/valory/contracts/multisend/contract.py
@@ -30,7 +30,7 @@ from aea.crypto.base import LedgerApi
 from hexbytes import HexBytes
 from web3 import Web3
 
-from packages.valory.contracts.constants import MIN_GAS, MIN_GASPRICE
+from packages.valory.contracts import MIN_GAS, MIN_GASPRICE
 
 
 PUBLIC_ID = PublicId.from_str("valory/multisend:0.1.0")
